Silenced 64-bit compiler warnings (loss of precision when assigning a 64-bit value to a 32-bit variable or parameter).

+++ src/e_basis.cc
@@ -728,7 +728,7 @@ public:
 
 	~undo_group_c()
 	{
-		for (int i = ops.size()-1; i >= 0; i--)
+		for (int i = (int)ops.size() - 1 ; i >= 0 ; i--)
 		{
 			ops[i].Destroy();
 		}
